Reorganize some preprocessor related code.
